Switch layout to be a client layout again
